refactor: simplify LatestProposalsTable by integrating ProposalsTableWithPagination
