+++ src/app/proposals/components/LatestProposalsTable.tsx
@@ -1,38 +1,21 @@
 'use client'
 import { useMemo, memo, useState, useRef, useCallback, useEffect } from 'react'
 import { AnimatePresence, motion } from 'motion/react'
-import {
-  createColumnHelper,
-  type SortingState,
-  getCoreRowModel,
-  getSortedRowModel,
-  useReactTable,
-  getPaginationRowModel,
-  PaginationState,
-} from '@tanstack/react-table'
-import Big from '@/lib/big'
-import { GridTable } from '@/components/Table'
-import { ProposalNameColumn, ProposerColumn } from './table-columns/ProposalNameColumn'
-import { QuorumColumn, VotesColumn } from './table-columns/VotesColumn'
-import { Countdown } from '@/components/Countdown'
 import { DebounceSearch } from '@/components/DebounceSearch'
-import { useSearchParams } from 'next/navigation'
 import { FilterButton } from './filter/FilterButton'
 import { FilterSideBar } from './filter/FilterSideBar'
 import { cn } from '@/lib/utils'
 import { useClickOutside } from '@/shared/hooks/useClickOutside'
-import { Status } from '@/components/Status'
 import { SearchButton } from './SearchButton'
-import { Category } from '../components/category'
-import { Header, Paragraph } from '@/components/Typography'
+import { Header } from '@/components/Typography'
 import { Proposal } from '@/app/proposals/shared/types'
-import { Pagination } from '@/components/Pagination'
+import { filterOptions } from './filter/filterOptions'
 import { useIsDesktop } from '@/shared/hooks/useIsDesktop'
 import { useStickyHeader } from '@/shared/hooks'
 import { ActiveFiltersDisplay } from './filter/ActiveFiltersDisplay'
 import { useProposalFilters } from './filter/useProposalFilters'
 import { FilterType } from './filter/types'
-import { filterOptions } from './filter/filterOptions'
+import { ProposalsTable, type ProposalsTableRef } from './ProposalsTableWithPagination'
 
 interface LatestProposalsTableProps {
   proposals: Proposal[]
@@ -40,23 +23,13 @@ interface LatestProposalsTableProps {
 
 const LatestProposalsTable = ({ proposals }: LatestProposalsTableProps) => {
   const isDesktop = useIsDesktop()
-  // React-table sorting state
-  const [sorting, setSorting] = useState<SortingState>([])
+  const proposalsTableRef = useRef<ProposalsTableRef>(null)
 
   // Sticky header hook - only enabled on mobile/tablet
   const { headerRef } = useStickyHeader({
     isEnabled: !isDesktop,
     backgroundColor: 'var(--color-bg-80)',
   })
-
-  const searchParams = useSearchParams()
-
-  // Convert 1-indexed URL page to 0-indexed internal page
-  const [pagination, setPagination] = useState<PaginationState>(() => ({
-    pageIndex: Math.max(parseInt(searchParams?.get('page') ?? '1') - 1, 0),
-    pageSize: 10,
-  }))
-  const resetPagination = useCallback(() => setPagination(prev => ({ ...prev, pageIndex: 0 })), [])
 
   // Enhanced filtering system
   const { activeFilters, searchValue, addFilter, removeFilter, clearAllFilters, updateSearchValue } =
@@ -71,9 +44,9 @@ const LatestProposalsTable = ({ proposals }: LatestProposalsTableProps) => {
   const handleSearch = useCallback(
     (value: string) => {
       updateSearchValue(value)
-      resetPagination()
+      proposalsTableRef.current?.resetPagination()
     },
-    [updateSearchValue, resetPagination],
+    [updateSearchValue],
   )
 
   // show searchfield focus on SearchButton click
@@ -111,160 +84,6 @@ const LatestProposalsTable = ({ proposals }: LatestProposalsTableProps) => {
   const hasSelectedFilters = useMemo(() => {
     return activeFilters.filter(f => !f.isAll && f.type !== FilterType.SEARCH).length > 0
   }, [activeFilters])
-
-  // Table data definition helper
-  const { accessor } = createColumnHelper<(typeof proposals)[number]>()
-  // Table columns definition
-  const columns = [
-    accessor('name', {
-      id: 'name',
-      cell: ({ cell, row }) => (
-        <ProposalNameColumn name={cell.getValue()} proposalId={row.original.proposalId} />
-      ),
-    }),
-    accessor('proposer', {
-      id: 'proposer',
-      header: 'Proposal name',
-      sortDescFirst: false,
-      cell: ({ cell }) => <ProposerColumn by={cell.getValue()} />,
-      sortingFn: (a, b) => a.original.name.localeCompare(b.original.name),
-      meta: {
-        width: '1.6fr',
-      },
-    }),
-    accessor('Starts', {
-      id: 'date',
-      header: 'Date',
-      sortDescFirst: false,
-      cell: ({ cell }) => <Paragraph>{cell.getValue().format('MMM DD, YYYY')}</Paragraph>,
-      meta: {
-        width: '1.2fr',
-      },
-    }),
-    accessor('blocksUntilClosure', {
-      id: 'timeRemaining',
-      header: 'Vote ending in',
-      sortDescFirst: false,
-      cell: info => {
-        const { proposalDeadline, blockNumber } = info.row.original
-        return (
-          <Countdown
-            end={proposalDeadline}
-            timeSource="blocks"
-            referenceStart={Big(blockNumber)}
-            colorDirection="normal"
-          />
-        )
-      },
-      sortingFn: (a, b) => {
-        return a.original.blocksUntilClosure.cmp(b.original.blocksUntilClosure)
-      },
-      meta: {
-        width: '1.32fr',
-      },
-    }),
-    accessor('votes', {
-      id: 'quorum',
-      sortDescFirst: false,
-      header: () => (
-        <div>
-          <div className="flex items-center-safe gap-1">
-            <p className="mb-1">Quorum</p>
-          </div>
-          <Paragraph variant="body-xs" className="text-text-40">
-            {isFilterSidebarOpen ? 'reached' : 'needed | reached'}
-          </Paragraph>
-        </div>
-      ),
-      cell: ({ row }) => {
-        const quorum = row.original.quorumAtSnapshot
-        const { forVotes, abstainVotes } = row.original.votes
-        return (
-          <QuorumColumn
-            quorumVotes={forVotes.add(abstainVotes)}
-            quorumAtSnapshot={quorum}
-            hideQuorumTarget={isFilterSidebarOpen}
-          />
-        )
-      },
-      sortingFn: (a, b) => {
-        const getQuorumPercent = (quorum: Big, quorumAtSnapshot: Big) =>
-          quorumAtSnapshot.eq(0)
-            ? Big(0)
-            : quorum.div(quorumAtSnapshot).mul(100).round(undefined, Big.roundHalfEven)
-        const percentA = getQuorumPercent(a.original.votes.quorum, a.original.quorumAtSnapshot)
-        const percentB = getQuorumPercent(b.original.votes.quorum, b.original.quorumAtSnapshot)
-        return percentA.cmp(percentB)
-      },
-      meta: {
-        width: '1.4fr',
-      },
-    }),
-    accessor('votes', {
-      id: 'votes',
-      header: 'Votes',
-      sortDescFirst: false,
-      cell: ({ cell }) => {
-        const { forVotes, abstainVotes, againstVotes } = cell.getValue()
-        return (
-          <VotesColumn
-            forVotes={forVotes.toNumber()}
-            againstVotes={againstVotes.toNumber()}
-            abstainVotes={abstainVotes.toNumber()}
-          />
-        )
-      },
-      sortingFn: (a, b) => {
-        const votesA = a.original.votes
-        const votesB = b.original.votes
-        const sumA = votesA.forVotes.add(votesA.againstVotes).add(votesA.abstainVotes)
-        const sumB = votesB.forVotes.add(votesB.againstVotes).add(votesB.abstainVotes)
-        return sumA.cmp(sumB)
-      },
-      meta: {
-        width: '1fr',
-      },
-    }),
-    accessor('category', {
-      id: 'propType',
-      header: 'Type',
-      sortDescFirst: false,
-      meta: {
-        width: '0.62fr',
-      },
-      cell: ({ cell }) => <Category category={cell.getValue()} />,
-    }),
-    accessor('proposalState', {
-      id: 'status',
-      header: 'Status',
-      sortDescFirst: false,
-      cell: ({ cell }) => (
-        <div className="w-full flex justify-center">
-          <Status proposalState={cell.getValue()} />
-        </div>
-      ),
-      meta: {
-        width: '0.8fr',
-      },
-    }),
-  ]
-
-  // create table data model which is passed to the Table UI component
-  const table = useReactTable({
-    columns,
-    data: filteredProposalList,
-    state: {
-      sorting,
-      pagination,
-    },
-    onSortingChange: setSorting,
-    getCoreRowModel: getCoreRowModel(),
-    getSortedRowModel: getSortedRowModel(),
-    getPaginationRowModel: getPaginationRowModel(),
-    onPaginationChange: setPagination, //update the pagination state when internal APIs mutate the pagination state
-    // Prevent pagination reset on data change
-    autoResetPageIndex: false,
-  })
 
   return (
     <div className="py-4 px-6 rounded-sm bg-bg-80">
@@ -341,21 +160,11 @@ const LatestProposalsTable = ({ proposals }: LatestProposalsTableProps) => {
         </motion.div>
         <div className="grow overflow-y-auto">
           {filteredProposalList.length > 0 ? (
-            <div>
-              <GridTable
-                className="min-w-[600px]"
-                aria-label="Proposal table"
-                stackFirstColumn
-                table={table}
-                data-testid="TableProposals"
-              />
-              <Pagination
-                pagination={pagination}
-                setPagination={setPagination}
-                data={filteredProposalList}
-                table={table}
-              />
-            </div>
+            <ProposalsTable
+              ref={proposalsTableRef}
+              proposals={filteredProposalList}
+              isFilterSidebarOpen={isFilterSidebarOpen}
+            />
           ) : (
             <p data-testid="NoProposals">No proposals found &#x1F622;</p>
           )}
